+++ CHANGELOG.md
@@ -22,7 +22,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **ID3v2**: `LanguageFrame`'s `lang` field has changed type - `String` -> `[u8; 3]`
 - **MP3**: Renamed `MP3File` -> `MPEGFile`
 - **MP4**: `Atom::data` will now return all values
-- **Vorbis Comments**: Recognize lowercase `METADATA_BLOCK_PICTURE` as a picture
+- **Vorbis Comments**: Recognize lowercase `METADATA_BLOCK_PICTURE` as a picture ([issue](https://github.com/Serial-ATA/lofty-rs/issues/60))
+
+### Fixed
+- **ID3v2**: `ItemKey::InitialKey` now maps to `TKEY` ([PR](https://github.com/Serial-ATA/lofty-rs/pull/61))
 
 ## [0.7.3] - 2022-07-22